Improved initial loading speed for chat rooms.

+++ app/component/ProfileImageDisplay.ts
@@ -1,13 +1,26 @@
-import { DomNode, el } from "common-dapp-module";
+import { DomNode, ErrorAlert, el } from "common-dapp-module";
 import { generateJazziconDataURL } from "common-dapp-module/lib/component/Jazzicon.js";
 import OnlineUserManager from "../OnlineUserManager.js";
 import UserDetailsCacher from "../cacher/UserDetailsCacher.js";
+import UserDetails from "../data/UserDetails.js";
+import UserInfoPopup from "../popup/user/UserInfoPopup.js";
 
 export default class ProfileImageDisplay extends DomNode {
   private currentWalletAddress: string | undefined;
+  private currentUserDetails: UserDetails | undefined;
 
   constructor(private isLarge = false) {
-    super(".profile-image-display.loading");
+    super("a.profile-image-display.loading");
+    this.onDom("click", () => {
+      if (this.currentUserDetails) {
+        new UserInfoPopup(this.currentUserDetails);
+      } else {
+        new ErrorAlert({
+          title: "User not found",
+          message: "User not found",
+        });
+      }
+    });
     this.onDelegate(
       OnlineUserManager,
       "onlineUsersChanged",
@@ -31,6 +44,8 @@ export default class ProfileImageDisplay extends DomNode {
     this.checkOnline();
 
     const userDetails = await UserDetailsCacher.get(walletAddress);
+    this.currentUserDetails = userDetails;
+
     if (userDetails) {
       this.src = this.isLarge
         ? userDetails.profile_image.replace("_normal", "")
@@ -43,8 +58,10 @@ export default class ProfileImageDisplay extends DomNode {
   private set src(src: string) {
     const img = el<HTMLImageElement>("img");
     img.onDom("load", () => {
-      this.empty().append(img);
-      this.deleteClass("loading");
+      if (!this.deleted) {
+        this.empty().append(img);
+        this.deleteClass("loading");
+      }
     });
     img.domElement.src = src;
   }
